+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - **copilot:** copilot now not loaded if there's no config
 - **copilot:** disabled copilot lua if copilot vim is not initialized yet
 - **hls-lens:** moved mapping to lua
+- **settings:** update setting to use nvim_cmd instead of vim string
 - **theme:** mvoed to one nord
 
 > Revert